requirements/base.txt
Django==4.0
djangorestframework==3.13.1
django-cors-headers==3.13.0
djangorestframework-simplejwt==5.2.0
django-ckeditor==6.4.2
drf-spectacular==0.22.1
django-filter==22.1
django-autoslug==1.9.8
Pillow==9.2.0
django-admin-inline-paginator==0.3.0